\name{as.SarimaModel}
\alias{as.SarimaModel}
\alias{as.SarimaModel.Arima}

\title{Convert S3 model objects to class SarimaModel}
\description{Convert S3 model objects to class SarimaModel.}
\usage{
as.SarimaModel(x, \dots)

\method{as.SarimaModel}{Arima}(x, \dots)
}
\arguments{
  \item{x}{an objects from a class representing Seasonal ARIMA models.}
  \item{\dots}{further arguments for methods.}
}
\details{

  This function can be useful when one needs to manipulate the
  components of SARIMA models.
  
  The method for class \code{Arima} (objects generated by \code{stats::arima()})
  extracts the model information and convert it to \code{"SarimaModel"}.

  For S4 classes, there are methods for \code{as()}, where
  suitable. \code{\link{modelCoef}} provides a more powerful conversion
  mechanism.

}
\value{
  an object from class \code{"SarimaModel"}
}
%\references{
%%% ~put references to the literature/web site here ~
%}
%\author{
%%%  ~~who you are~~
%}
%\note{
%  I made this function because I didn't want to convert class
%  \code{"Arima"} to S4.
%}
%
%%% ~Make other sections like Warning with \section{Warning }{....} ~
%
\seealso{
  \code{\linkS4class{SarimaModel}}
}
%\examples{
%}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.




