\name{autocorrelations-methods}
\docType{methods}
\alias{autocorrelations-methods}
\alias{autocorrelations,ANY,ANY,ANY-method}
\alias{autocorrelations,ANY,ANY,missing-method}
\alias{autocorrelations,Autocorrelations,ANY,missing-method}
\alias{autocorrelations,Autocorrelations,missing,missing-method}
\alias{autocorrelations,Autocovariances,ANY,missing-method}
\alias{autocorrelations,PartialAutocorrelations,ANY,missing-method}
\alias{autocorrelations,PartialAutocovariances,ANY,missing-method}
\alias{autocorrelations,SamplePartialAutocorrelations,ANY,missing-method}
\alias{autocorrelations,SamplePartialAutocovariances,ANY,missing-method}
\alias{autocorrelations,VirtualArmaModel,ANY,missing-method}
\alias{autocorrelations,VirtualSarimaModel,ANY,missing-method}
\title{Methods for function autocorrelations()}
\description{Methods for function autocorrelations().}
\author{Georgi N. Boshnakov}

\section{Methods}{
\describe{

\item{\code{signature(x = "ANY", maxlag = "ANY", lag_0 = "ANY")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(x = "ANY", maxlag = "ANY", lag_0 = "missing")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(x = "Autocorrelations", maxlag = "ANY", lag_0 = "missing")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(x = "Autocorrelations", maxlag = "missing", lag_0 = "missing")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(x = "Autocovariances", maxlag = "ANY", lag_0 = "missing")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(x = "PartialAutocorrelations", maxlag = "ANY", lag_0 = "missing")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(x = "PartialAutocovariances", maxlag = "ANY", lag_0 = "missing")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(x = "SamplePartialAutocorrelations", maxlag = "ANY", lag_0 = "missing")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(x = "SamplePartialAutocovariances", maxlag = "ANY", lag_0 = "missing")}}{
%%  ~~describe this method here~~
}
\item{\code{signature(x = "VirtualArmaModel", maxlag = "ANY", lag_0 = "missing")}}{{
%%  ~~describe this method here~~
}
}\item{\code{signature(x = "VirtualSarimaModel", maxlag = "ANY", lag_0 = "missing")}}{{
%%  ~~describe this method here~~
}
}}}
\examples{
## see the examples for ?autocorrelations
}
\keyword{methods}
