\name{filterOrder-methods}
\docType{methods}
\alias{filterOrder-methods}
\alias{filterOrder,MonicFilterSpec-method}
\alias{filterOrder,SarimaFilter-method}
\alias{filterOrder,VirtualArmaFilter-method}
\title{Methods for filterOrder()}
\description{Methods for filterOrder().}
\section{Methods}{
\describe{

\item{\code{signature(object = "MonicFilterSpec")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "SarimaFilter")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "VirtualArmaFilter")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
\keyword{ ~~ other possible keyword(s) ~~ }
