\name{filterPolyCoef-methods}
\docType{methods}
\alias{filterPolyCoef-methods}
\alias{filterPolyCoef,BJFilter-method}
\alias{filterPolyCoef,SarimaFilter-method}
\alias{filterPolyCoef,SPFilter-method}
\alias{filterPolyCoef,VirtualArmaFilter-method}
\alias{filterPolyCoef,VirtualBJFilter-method}
\alias{filterPolyCoef,VirtualSPFilter-method}
\title{Methods for filterPolyCoef}
\description{Methods for filterPolyCoef.}
\section{Methods}{
\describe{

\item{\code{signature(object = "BJFilter")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "SarimaFilter")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "SPFilter")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "VirtualArmaFilter")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "VirtualBJFilter")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "VirtualSPFilter")}}{
%%  ~~describe this method here~~
}
}}
\author{Georgi N. Boshnakov}
\keyword{methods}
