\name{modelCoef-methods}
\docType{methods}
\alias{modelCoef-methods}
\alias{modelCoef,Autocorrelations,ComboAutocorrelations-method}
\alias{modelCoef,Autocorrelations,PartialAutocorrelations-method}
\alias{modelCoef,Autocovariances,Autocorrelations-method}
\alias{modelCoef,Autocovariances,ComboAutocorrelations-method}
\alias{modelCoef,Autocovariances,ComboAutocovariances-method}
\alias{modelCoef,Autocovariances,PartialAutocorrelations-method}
\alias{modelCoef,ComboAutocorrelations,Autocorrelations-method}
\alias{modelCoef,ComboAutocorrelations,PartialAutocorrelations-method}
\alias{modelCoef,ComboAutocovariances,Autocovariances-method}
\alias{modelCoef,ComboAutocovariances,PartialAutocovariances-method}
\alias{modelCoef,ComboAutocovariances,PartialVariances-method}
\alias{modelCoef,ComboAutocovariances,VirtualAutocovariances-method}
\alias{modelCoef,PartialAutocorrelations,Autocorrelations-method}
\alias{modelCoef,SarimaModel,ArFilter-method}
\alias{modelCoef,SarimaModel,ArmaFilter-method}
\alias{modelCoef,SarimaModel,character-method}
\alias{modelCoef,SarimaModel,MaFilter-method}
\alias{modelCoef,SarimaModel,missing-method}
\alias{modelCoef,SarimaModel,SarimaFilter-method}
\alias{modelCoef,VirtualArmaModel,character-method}
\alias{modelCoef,VirtualArmaModel,missing-method}
\alias{modelCoef,VirtualAutocovariances,character-method}
\alias{modelCoef,VirtualAutocovariances,missing-method}
\alias{modelCoef,VirtualAutocovariances,VirtualAutocovariances-method}
\alias{modelCoef,PartialAutocovariances,PartialAutocorrelations-method}
\title{Methods for generic function modelCoef}
\description{Methods for generic function modelCoef.}
\section{Methods}{
\describe{

\item{\code{signature(object = "Autocorrelations", convention = "ComboAutocorrelations")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "Autocorrelations", convention = "PartialAutocorrelations")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "Autocovariances", convention = "Autocorrelations")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "Autocovariances", convention = "ComboAutocorrelations")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "Autocovariances", convention = "ComboAutocovariances")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "Autocovariances", convention = "PartialAutocorrelations")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "ComboAutocorrelations", convention = "Autocorrelations")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "ComboAutocorrelations", convention = "PartialAutocorrelations")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "ComboAutocovariances", convention = "Autocovariances")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "ComboAutocovariances", convention = "PartialAutocovariances")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "ComboAutocovariances", convention = "PartialVariances")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "ComboAutocovariances", convention = "VirtualAutocovariances")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "PartialAutocorrelations", convention = "Autocorrelations")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "SarimaModel", convention = "ArFilter")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "SarimaModel", convention = "ArmaFilter")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "SarimaModel", convention = "character")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "SarimaModel", convention = "MaFilter")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "SarimaModel", convention = "missing")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "SarimaModel", convention = "SarimaFilter")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "VirtualArmaModel", convention = "character")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "VirtualArmaModel", convention = "missing")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "VirtualAutocovariances", convention = "character")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "VirtualAutocovariances", convention = "missing")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "VirtualAutocovariances", convention = "VirtualAutocovariances")}}{
%%  ~~describe this method here~~
}
\item{\code{signature(object = "PartialAutocovariances", 
          convention = "PartialAutocorrelations")}}{{
%%  ~~describe this method here~~
}
}}}
\keyword{methods}
%\keyword{}
