\name{modelOrder-methods}
\docType{methods}
\alias{modelOrder-methods}
\alias{modelOrder,ArmaModel,ArFilter-method}
\alias{modelOrder,ArmaModel,MaFilter-method}
%\alias{modelOrder,ArmaModel,missing-method}
\alias{modelOrder,SarimaModel,ArFilter-method}
\alias{modelOrder,SarimaModel,ArmaFilter-method}
\alias{modelOrder,SarimaModel,ArmaModel-method}
\alias{modelOrder,SarimaModel,ArModel-method}
\alias{modelOrder,SarimaModel,MaFilter-method}
\alias{modelOrder,SarimaModel,MaModel-method}
%\alias{modelOrder,SarimaModel,missing-method}
\alias{modelOrder,VirtualFilterModel,missing-method}
\title{Get the order of a model}
\description{Get the order of a model.}
\section{Methods}{
\describe{

\item{\code{signature(object = "ArmaModel", convention = "ArFilter")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "ArmaModel", convention = "MaFilter")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "ArmaModel", convention = "missing")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "SarimaModel", convention = "ArFilter")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "SarimaModel", convention = "ArmaFilter")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "SarimaModel", convention = "ArmaModel")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "SarimaModel", convention = "ArModel")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "SarimaModel", convention = "MaFilter")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "SarimaModel", convention = "MaModel")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "SarimaModel", convention = "missing")}}{
%%  ~~describe this method here~~
}
\item{\code{signature(object = "VirtualFilterModel", convention = "missing")}}{{
%%  ~~describe this method here~~
}
}}}
\author{Georgi N. Boshnakov}
\keyword{methods}
