\name{modelPoly-methods}
\docType{methods}
\alias{modelPoly-methods}
\alias{modelPoly,SarimaModel,ArmaFilter-method}
\alias{modelPoly,SarimaModel,missing-method}
\title{Get polynomials associated with SARIMA models}
\description{Get polynomials associated with SARIMA models.}
\section{Methods}{
\describe{

\item{\code{signature(object = "SarimaModel", convention = "ArmaFilter")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "SarimaModel", convention = "missing")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
\keyword{sarima}
