\name{modelPolyCoef-methods}
\docType{methods}
\alias{modelPolyCoef-methods}
\alias{modelPolyCoef,SarimaModel,ArmaFilter-method}
\alias{modelPolyCoef,SarimaModel,missing-method}
\title{Methods for modelPolyCoef}
\description{Methods for modelPolyCoef, e generic function for getting
  the coefficients of polynomials associated with SARIMA models.}
\section{Methods}{
\describe{

\item{\code{signature(object = "SarimaModel", convention = "ArmaFilter")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "SarimaModel", convention = "missing")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
\keyword{ ~~ other possible keyword(s) ~~ }
