\name{partialAutocorrelations-methods}
\docType{methods}
\alias{partialAutocorrelations-methods}
\alias{partialAutocorrelations,ANY,ANY,ANY-method}
\alias{partialAutocorrelations,mts,ANY,missing-method}
\alias{partialAutocorrelations,PartialAutocovariances,ANY,missing-method}
\alias{partialAutocorrelations,ts,ANY,missing-method}
\title{Methods for function partialAutocorrelations}
\description{Methods for function partialAutocorrelations.}
\section{Methods}{
\describe{

\item{\code{signature(x = "ANY", maxlag = "ANY", lag_0 = "ANY")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(x = "mts", maxlag = "ANY", lag_0 = "missing")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(x = "PartialAutocovariances", maxlag = "ANY", lag_0 = "missing")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(x = "ts", maxlag = "ANY", lag_0 = "missing")}}{
%%  ~~describe this method here~~
}
}}
\author{Georgi N. Boshnakov}
\keyword{methods}
