% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarimaFormulae.R
\name{sarimaFormulae}
\alias{sarimaFormulae}
\alias{ar}
\alias{ma}
\alias{sar}
\alias{sma}
\alias{i}
\alias{s}
\alias{u}
\alias{si}
\alias{ss}
\alias{su}
\alias{uar}
\title{Model formulae for SARIMA models}
\usage{
ar(p, ar, sign = "-", atanh.tr = TRUE)
  ma(q, ma, sign = "+", atanh.tr = TRUE)
  sar(s, p, ar, sign = "-", atanh.tr = TRUE)
  sma(s, q, ma, sign = "+", atanh.tr = TRUE)
  i(d)
  s(...)
  u(u, fixed = TRUE, operator = all(fixed))
  uar(p, parcor, sign = "-", atanh.tr = TRUE, fixed = NULL)
  si(s, d)
  ss(s, ...)
  su(u, h)
}
\arguments{
\item{p}{A positive integer value for the degree of the autoregressive polynomial.}

\item{q}{A positive integer value for the degree of the moving average polynomial.}

\item{d}{A positive integer value for the degree of the difference polynomial.}

\item{s}{A positive integer value for the degree of the seasonal effects.}

\item{h}{A positive integer value specfying the harmonics of the seasonal unit root 
polynomial.}

\item{ar}{A numeric vector containing initial values for the coefficients of 
the autoregressive polynomial.}

\item{ma}{A numeric vector containing initial values for the coefficients of 
the moving average polynomial.}

\item{u}{If real, the argument of one of a pair of complex unit roots as 
a fraction of \eqn{2\pi}. If complex, it is the root itself.}

\item{parcor}{A numeric vector containing initial values for the partial 
autocoefficients of the time series.}

\item{sign}{A character argument indicating the sign to be used for the 
coefficients in the polynomial. Should be one of "-" or "+".}

\item{atanh.tr}{A logical indicator for whether the partial autocorrelations 
should be transformed by arctanh().}

\item{operator}{An indicator value for unit root terms, where TRUE indicates
all fixed root and FALSE indicates that at least one coefficient must be
estimated.}

\item{fixed}{Either a logical vector of length one or of the same length as 
the coefficeints indicating which values are fixed. Alternatively, it can
be a vector of positive integer values specifying the fixed coefficients.}

\item{...}{Further arguments. In \code{s} and \code{ss} these should be 
positive integer values indicating the number of terms to be summed.}
}
\description{
A flexible syntax is provided for the specification of the SARIMA part of the 
   model. It is formed using a number of primitives for stationary and unit 
   root components, which have non-seasonal and seasonal variants. These are 
   explained in detail here. Arbitrary number of multiplicative factors and 
   multiple seasonalities can be specified.
}
\details{
\describe{
  \item{ar(p)}{autoregression term of order p}
  \item{ma(q)}{moving average term of order q}
  \item{sar(s,p)}{seasonal autoregression term (s seasons, order p)}
  \item{sma(s,q)}{seasonal moving average term (s seasons, order q)}
  \item{i(d)}{difference term, \eqn{(1-B)^d}}
  \item{s(c)}{summation operator, \eqn{(1 + B + \dots + B^{c-1})}}
  \item{u(x)}{quadratic unit root term, corresponding to a complex pair on the 
        unit circle. If \eqn{x} is real, it specifies the argument of one of the 
        roots as a fraction of \eqn{2\pi}. If \eqn{z} is complex, it is the 
        root itself. The real roots of modulus one (1 and \eqn{-1}) should be 
        specified using \code{i(1)} and \code{s(2)}, which correspond to 
        \eqn{1-B} and \eqn{1+B}, respectively.}
  \item{uar(p)}{unit root polynomial term (order p)}
  \item{si(s, d)}{seasonal difference term, \eqn{(1-B^s)^d}}
  \item{su(s, h)}{quadratic unit root terms corresponding to seasonal 
        differencing factors. h specifies the desired harmonic which should be 
        one of 1,2, ..., [s/2]. Several harmonics can be specified by setting h 
        to a vector.}    
  \item{ss(s, p)}{seasonal summation operator, \eqn{(1 + B^s + \dots + B^{(s-1)p})}}
}
}
\seealso{
\code{\link{sarima}}
}
