\name{VirtualMonicFilter-class}
\Rdversion{1.1}
\docType{class}

\alias{BJ-class}
\alias{SP-class}
\alias{BD-class}
%\alias{modelCoef,VirtualFilterModel,BJ-method}
%\alias{modelCoef,VirtualFilterModel,SP-method}
%\alias{modelCoef,VirtualFilterModel,BD-method}

\alias{VirtualMonicFilter-class}
\alias{VirtualBJFilter-class}
\alias{VirtualSPFilter-class}
\alias{VirtualArmaFilter-class}

\alias{VirtualAutocovarianceModel-class}
\alias{VirtualAutocorelationModel-class}
\alias{VirtualPartialAutocovarianceModel-class}
\alias{VirtualPartialAutocorelationModel-class}

\alias{VirtualIntegratedModel-class}
\alias{VirtualSarimaModel-class}
\alias{VirtualImaModel-class}

\alias{VirtualFilterModel-class}
\alias{VirtualArmaModel-class}

\alias{VirtualAutocovarianceSpec-class}
\alias{VirtualMonicFilterSpec-class}

\alias{VirtualWhiteNoiseModel-class}

\alias{VirtualArModel-class}
\alias{VirtualMaModel-class}

\alias{VirtualArimaModel-class}
\alias{VirtualMeanModel-class}
\alias{VirtualAriModel-class}
\alias{VirtualSarimaFilter-class}
\alias{VirtualStationaryModel-class}
\alias{VirtualCascadeFilter-class}

\alias{MonicFilterSpec-class}
\alias{BJFilter-class}
\alias{SPFilter-class}
\alias{ArmaFilter-class}
\alias{ArFilter-class}
\alias{MaFilter-class}
\alias{SarimaFilter-class}

%\alias{ArmaModel-class}
%\alias{ArModel-class}
%\alias{MaModel-class}

\alias{SarimaSpec-class}
%\alias{SarimaModel-class}

\alias{AutocovarianceSpec-class}
\alias{AutocovarianceModel-class}

\alias{ArmaSpec-class}


\alias{VirtualAutocorrelations-class}
\alias{VirtualAutocovariances-class}
\alias{Fitted-class}

\alias{Autocovariances-class}
\alias{Autocorrelations-class}


\alias{ComboAutocovariances-class}
\alias{ComboAutocorrelations-class}

\alias{PartialAutocovariances-class}
\alias{PartialAutocorrelations-class}
\alias{PartialVariances-class}

\alias{SampleAutocovariances-class}
\alias{SampleAutocorrelations-class}
\alias{SamplePartialAutocovariances-class}
\alias{SamplePartialAutocorrelations-class}
\alias{SamplePartialVariances-class}


\title{Undocumented classes in package sarima}
\description{This page is for classes without proper documentation.}
\section{Objects from the Class}{A virtual Class: No objects may be
  created from it.

  This page exists only to remind me which classes do not have
  documentation yet. It exists to avoid cluttering the report from 'R
  CMD check' during early stages of development.

}
\section{Methods}{
No methods defined with class "VirtualMonicFilter" in the signature.
}
%\references{
%%%  ~~put references to the literature/web site here~~
%}
\author{Georgi N. Boshnakov}
%\note{
%%%  ~~further notes~~
%}
%
%%% ~Make other sections like Warning with \section{Warning }{....} ~
%
%\seealso{
%%%  ~~objects to See Also as \code{\link{~~fun~~}}, ~~~
%%%  ~~or \code{\linkS4class{CLASSNAME}} for links to other classes ~~~
%}
%\examples{
%showClass("VirtualMonicFilter")
%}
\keyword{classes}
\keyword{internal}
