\name{autocovariances-methods}
\docType{methods}
\alias{autocovariances-methods}
\alias{autocovariances,ANY-method}
\alias{autocovariances,VirtualArmaModel-method}
\title{Methods for function autocovariances()}
\description{Methods for function autocovariances().}
\author{Georgi N. Boshnakov}
\section{Methods}{
\describe{

\item{\code{signature(x = "ANY")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(x = "VirtualArmaModel")}}{
%%  ~~describe this method here~~
}
}}
\seealso{
  \code{\link{autocorrelations}}
}
\examples{
## see the examples for ?autocorrelations
}
\keyword{methods}
