\name{autocovariances-methods}
\docType{methods}
\alias{autocovariances-methods}

\alias{autocovariances,ANY,ANY-method}
\alias{autocovariances,Autocovariances,ANY-method}
\alias{autocovariances,Autocovariances,missing-method}
\alias{autocovariances,VirtualArmaModel,ANY-method}
\alias{autocovariances,VirtualAutocovariances,ANY-method}
\title{Methods for function autocovariances()}
\description{Methods for function autocovariances().}
\author{Georgi N. Boshnakov}
\section{Methods}{
\describe{

\item{\code{signature(x = "ANY", maxlag = "ANY")}}{{
%%  ~~describe this method here~~
}
}\item{\code{signature(x = "Autocovariances", maxlag = "ANY")}}{{
%%  ~~describe this method here~~
}
}\item{\code{signature(x = "Autocovariances", maxlag = "missing")}}{{
%%  ~~describe this method here~~
}
}\item{\code{signature(x = "VirtualArmaModel", maxlag = "ANY")}}{{
%%  ~~describe this method here~~
}
}\item{\code{signature(x = "VirtualAutocovariances", maxlag = "ANY")}}{{
%%  ~~describe this method here~~
}
}}}
\seealso{
  \code{\link{autocorrelations}}
}
\examples{
## see the examples for ?autocorrelations
}
\keyword{methods}
