\name{filterCoef-methods}
\docType{methods}
\alias{filterCoef-methods}
\alias{filterCoef,BJFilter,character-method}
%\alias{filterCoef,MonicFilterSpec,missing-method}
\alias{filterCoef,SPFilter,character-method}
\alias{filterCoef,VirtualArmaFilter,character-method}
\alias{filterCoef,VirtualArmaFilter,missing-method}
\alias{filterCoef,VirtualBJFilter,character-method}
\alias{filterCoef,VirtualSPFilter,character-method}
\alias{filterCoef,SarimaFilter,ANY-method}
\alias{filterCoef,SarimaFilter,character-method}
\alias{filterCoef,VirtualMonicFilterSpec,missing-method}
\title{Methods for filterCoef()}
\description{Methods for filterCoef().}
\section{Methods}{
\describe{

\item{\code{signature(object = "BJFilter", convention = "character")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "SPFilter", convention = "character")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "VirtualArmaFilter", convention = "missing")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "VirtualArmaFilter", convention = "character")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "VirtualBJFilter", convention = "character")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "VirtualSPFilter", convention = "character")}}{
%%  ~~describe this method here~~
}
\item{\code{signature(object = "SarimaFilter", convention = "ANY")}}{{
%%  ~~describe this method here~~
}
}\item{\code{signature(object = "SarimaFilter", convention = "character")}}{{
%%  ~~describe this method here~~
}
}\item{\code{signature(object = "VirtualMonicFilterSpec", convention = "missing")}}{{
%%  ~~describe this method here~~
}
}}}
\author{Georgi N. Boshnakov}
\keyword{methods}
