\name{filterOrder-methods}
\docType{methods}
\alias{filterOrder-methods}
%\alias{filterOrder,MonicFilterSpec-method}
\alias{filterOrder,SarimaFilter-method}
\alias{filterOrder,VirtualArmaFilter-method}
\alias{filterOrder,VirtualMonicFilterSpec-method}
\title{Methods for filterOrder()}
\description{Methods for filterOrder().}
\section{Methods}{
\describe{

\item{\code{signature(object = "SarimaFilter")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "VirtualArmaFilter")}}{
%%  ~~describe this method here~~
}
\item{\code{signature(object = "VirtualMonicFilterSpec")}}{{
%%  ~~describe this method here~~
}
}}}
\author{Georgi N. Boshnakov}
\keyword{methods}
