\name{filterPoly-methods}
\docType{methods}
\alias{filterPoly-methods}
\alias{filterPoly,BJFilter-method}
\alias{filterPoly,SarimaFilter-method}
\alias{filterPoly,SPFilter-method}
\alias{filterPoly,VirtualArmaFilter-method}
\alias{filterPoly,VirtualMonicFilterSpec-method}
\title{Methods for filterPoly}
\description{Methods for filterPoly.}
\section{Methods}{
\describe{

\item{\code{signature(object = "BJFilter")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "SPFilter")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "SarimaFilter")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "VirtualArmaFilter")}}{
%%  ~~describe this method here~~
}
\item{\code{signature(object = "VirtualMonicFilterSpec")}}{{
%%  ~~describe this method here~~
}
}}}
\author{Georgi N. Boshnakov}
\keyword{methods}
