\name{modelCoef-methods}
\docType{methods}
\alias{modelCoef-methods}
\alias{modelCoef,Autocorrelations,ComboAutocorrelations,missing-method}
\alias{modelCoef,Autocorrelations,PartialAutocorrelations,missing-method}
\alias{modelCoef,Autocovariances,Autocorrelations,missing-method}
\alias{modelCoef,Autocovariances,ComboAutocorrelations,missing-method}
\alias{modelCoef,Autocovariances,ComboAutocovariances,missing-method}
\alias{modelCoef,Autocovariances,PartialAutocorrelations,missing-method}
\alias{modelCoef,ComboAutocorrelations,Autocorrelations,missing-method}
\alias{modelCoef,ComboAutocorrelations,PartialAutocorrelations,missing-method}
\alias{modelCoef,ComboAutocovariances,Autocovariances,missing-method}
\alias{modelCoef,ComboAutocovariances,PartialAutocovariances,missing-method}
\alias{modelCoef,ComboAutocovariances,PartialVariances,missing-method}
\alias{modelCoef,ComboAutocovariances,VirtualAutocovariances,missing-method}
\alias{modelCoef,PartialAutocorrelations,Autocorrelations,missing-method}
\alias{modelCoef,SarimaModel,ArFilter,missing-method}
\alias{modelCoef,SarimaModel,ArmaFilter,missing-method}
%\alias{modelCoef,SarimaModel,character,missing-method}
\alias{modelCoef,SarimaModel,MaFilter,missing-method}
%\alias{modelCoef,SarimaModel,missing,missing-method}
\alias{modelCoef,SarimaModel,SarimaFilter,missing-method}
%\alias{modelCoef,VirtualArmaModel,character,missing-method}
%\alias{modelCoef,VirtualArmaModel,missing,missing-method}
\alias{modelCoef,VirtualAutocovariances,character,missing-method}
\alias{modelCoef,VirtualAutocovariances,missing,missing-method}
\alias{modelCoef,VirtualAutocovariances,VirtualAutocovariances,missing-method}
\alias{modelCoef,PartialAutocovariances,PartialAutocorrelations,missing-method}
\alias{modelCoef,SarimaModel,ArModel,missing-method}
\alias{modelCoef,SarimaModel,MaModel,missing-method}
\alias{modelCoef,VirtualFilterModel,BD,missing-method}
\alias{modelCoef,VirtualFilterModel,BJ,missing-method}
\alias{modelCoef,VirtualFilterModel,character,missing-method}
\alias{modelCoef,VirtualFilterModel,missing,missing-method}
\alias{modelCoef,VirtualFilterModel,SP,missing-method}
\alias{modelCoef,ArmaModel,ArmaFilter,missing-method}
\alias{modelCoef,VirtualAutocovariances,Autocovariances,missing-method}
\title{Methods for generic function modelCoef}
\description{Methods for generic function modelCoef.}
\section{Methods}{
\describe{

\item{\code{signature(object = "Autocorrelations", convention = "ComboAutocorrelations", component = "missing")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "Autocorrelations", convention = "PartialAutocorrelations", component = "missing")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "Autocovariances", convention = "Autocorrelations", component = "missing")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "Autocovariances", convention = "ComboAutocorrelations", component = "missing")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "Autocovariances", convention = "ComboAutocovariances", component = "missing")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "Autocovariances", convention = "PartialAutocorrelations", component = "missing")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "ComboAutocorrelations", convention = "Autocorrelations", component = "missing")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "ComboAutocorrelations", convention = "PartialAutocorrelations", component = "missing")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "ComboAutocovariances", convention = "Autocovariances", component = "missing")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "ComboAutocovariances", convention = "PartialAutocovariances", component = "missing")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "ComboAutocovariances", convention = "PartialVariances", component = "missing")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "ComboAutocovariances", convention = "VirtualAutocovariances", component = "missing")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "PartialAutocorrelations", convention = "Autocorrelations", component = "missing")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "SarimaModel", convention = "ArFilter", component = "missing")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "SarimaModel", convention = "ArmaFilter", component = "missing")}}{
%%  ~~describe this method here~~
}


\item{\code{signature(object = "SarimaModel", convention = "MaFilter", component = "missing")}}{
%%  ~~describe this method here~~
}


\item{\code{signature(object = "SarimaModel", convention = "SarimaFilter", component = "missing")}}{
%%  ~~describe this method here~~
}


\item{\code{signature(object = "VirtualAutocovariances", convention = "character", component = "missing")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "VirtualAutocovariances", convention = "missing", component = "missing")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "VirtualAutocovariances", convention = "VirtualAutocovariances", component = "missing")}}{
%%  ~~describe this method here~~
}
\item{\code{signature(object = "SarimaModel", convention = "ArModel", component = "missing")}}{{
%%  ~~describe this method here~~
}
}\item{\code{signature(object = "SarimaModel", convention = "MaModel", component = "missing")}}{{
%%  ~~describe this method here~~
}
}\item{\code{signature(object = "VirtualFilterModel", convention = "BD", component = "missing")}}{{
%%  ~~describe this method here~~
}
}\item{\code{signature(object = "VirtualFilterModel", convention = "BJ", component = "missing")}}{{
%%  ~~describe this method here~~
}
}\item{\code{signature(object = "VirtualFilterModel", convention = "character", component = "missing")}}{{
%%  ~~describe this method here~~
}
}\item{\code{signature(object = "VirtualFilterModel", convention = "missing", component = "missing")}}{{
%%  ~~describe this method here~~
}
}\item{\code{signature(object = "VirtualFilterModel", convention = "SP", component = "missing")}}{{
%%  ~~describe this method here~~
}
}\item{\code{signature(object = "ArmaModel", convention = "ArmaFilter", 
          component = "missing")}}{{
%%  ~~describe this method here~~
}
}\item{\code{signature(object = "VirtualAutocovariances", 
          convention = "Autocovariances", component = "missing")}}{{
%%  ~~describe this method here~~
}
}}}
\author{Georgi N. Boshnakov}
\keyword{methods}
