\name{modelPoly-methods}
\docType{methods}
\alias{modelPoly-methods}
\alias{modelPoly,SarimaModel,ArmaFilter-method}
%\alias{modelPoly,SarimaModel,missing-method}
\alias{modelPoly,VirtualMonicFilter,missing-method}
\alias{modelPoly,VirtualFilterModel,character-method}
\title{Get polynomials associated with SARIMA models}
\description{Get polynomials associated with SARIMA models.}
\section{Methods}{
\describe{

\item{\code{signature(object = "SarimaModel", convention = "ArmaFilter")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(object = "VirtualMonicFilter", convention = "missing")}}{{
%%  ~~describe this method here~~
}
}\item{\code{signature(object = "VirtualFilterModel", convention = "character")}}{{
%%  ~~describe this method here~~
}
}}}
\author{Georgi N. Boshnakov}
\keyword{methods}
\keyword{sarima}
